comment module updated

+++ src/main/java/org/prashant/blog/blogapplicationapi/serviceimpl/CommentServiceImpl.java
@@ -9,14 +9,12 @@ import org.prashant.blog.blogapplicationapi.exceptions.UnAuthorizedOperationExcp
 import org.prashant.blog.blogapplicationapi.payload.*;
 import org.prashant.blog.blogapplicationapi.repository.CommentRepository;
 import org.prashant.blog.blogapplicationapi.repository.PostRepository;
-import org.prashant.blog.blogapplicationapi.repository.UserRepository;
 import org.prashant.blog.blogapplicationapi.service.CommentService;
 import org.prashant.blog.blogapplicationapi.service.UserService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 
 import java.util.Date;
@@ -27,16 +25,15 @@ import java.util.List;
 public class CommentServiceImpl implements CommentService {
     private final CommentRepository commentRepository;
     private final PostRepository postRepository;
-    private final UserRepository userRepository;
-    private  final UserService userService;
-    
+    private final UserService userService;
+
     @Override
     public CommentDTO createComment(CreateCommentRequest createCommentRequest) {
         Post post = postRepository.findById(createCommentRequest.postId())
-                .orElseThrow(()-> new ResourceNotFound("Post", "postId", createCommentRequest.postId().toString()));
+                .orElseThrow(() -> new ResourceNotFound("Post", "postId", createCommentRequest.postId().toString()));
 
         User user = userService.getLoggedInUser()
-                .orElseThrow(()-> new UnAuthorizedOperationExcpetion("Your are not authorized"));
+                .orElseThrow(() -> new UnAuthorizedOperationExcpetion("You are not authorized"));
 
         Comment comment = new Comment();
         comment.setCommentText(createCommentRequest.commentText());
@@ -45,56 +42,77 @@ public class CommentServiceImpl implements CommentService {
         comment.setCommentDate(new Date());
         comment.setLastEdited(new Date());
 
-        Comment saved_comment = commentRepository.save(comment);
-        return new CommentDTO(saved_comment);
+        Comment savedComment = commentRepository.save(comment);
+        return new CommentDTO(savedComment);
     }
 
     @Override
-    @PreAuthorize("#comment.getUser().getUsername() == authentication.name")
     public CommentDTO updateComment(UpdateCommentRequest request) {
         Comment comment = commentRepository.findById(request.commentId())
                 .orElseThrow(() -> new ResourceNotFound("Comment", "commentId", request.commentId().toString()));
+
+        User currentUser = userService.getLoggedInUser()
+                .orElseThrow(() -> new UnAuthorizedOperationExcpetion("You are not authorized"));
+
+        if (!comment.getUser().getUsername().equals(currentUser.getUsername())) {
+            throw new UnAuthorizedOperationExcpetion("You are not authorized to update this comment");
+        }
+
         comment.setCommentText(request.commentText());
+        comment.setLastEdited(new Date());
+
         Comment updatedComment = commentRepository.save(comment);
         return new CommentDTO(updatedComment);
     }
 
     @Override
-    @PreAuthorize("#comment.getUser().getUsername() == authentication.name")
     public void deleteComment(Long commentId) {
         Comment comment = commentRepository.findById(commentId)
-                .orElseThrow(()-> new ResourceNotFound("Comment", "commentId", commentId.toString()));
+                .orElseThrow(() -> new ResourceNotFound("Comment", "commentId", commentId.toString()));
+
+        User currentUser = userService.getLoggedInUser()
+                .orElseThrow(() -> new UnAuthorizedOperationExcpetion("You are not authorized"));
+
+        if (!comment.getUser().getUsername().equals(currentUser.getUsername())) {
+            throw new UnAuthorizedOperationExcpetion("You are not authorized to delete this comment");
+        }
+
         commentRepository.delete(comment);
     }
 
     @Override
     public CommentPageResponse getAllCommentsByPost(Long postId, Integer pageNumber, Integer pageSize, String sortBy, String sortDir) {
         Post post = postRepository.findById(postId)
-                .orElseThrow(()-> new ResourceNotFound("Post", "postId", postId.toString()));
+                .orElseThrow(() -> new ResourceNotFound("Post", "postId", postId.toString()));
+
         Sort sort = sortDir.equalsIgnoreCase("asc") ? Sort.by(sortBy).ascending() : Sort.by(sortBy).descending();
         Pageable pageable = PageRequest.of(pageNumber, pageSize, sort);
-        Page<Comment> page_comment = commentRepository.findByPost(post, pageable);
-        List<CommentDTO> all_comments = page_comment.getContent().stream().map(CommentDTO::new).toList();
-        return new CommentPageResponse(all_comments,
-                page_comment.getNumber(),
-                page_comment.getSize(),
-                page_comment.getTotalElements(),
-                page_comment.getTotalPages(), page_comment.isLast());
+        Page<Comment> pageComment = commentRepository.findByPost(post, pageable);
+
+        List<CommentDTO> allComments = pageComment.getContent().stream().map(CommentDTO::new).toList();
+        return new CommentPageResponse(allComments,
+                pageComment.getNumber(),
+                pageComment.getSize(),
+                pageComment.getTotalElements(),
+                pageComment.getTotalPages(),
+                pageComment.isLast());
     }
 
     @Override
     public CommentPageResponse getAllCommentsByUser(Integer pageNumber, Integer pageSize, String sortBy, String sortDir) {
         User user = userService.getLoggedInUser()
-                .orElseThrow(()-> new UnAuthorizedOperationExcpetion("Unauthorized Access to Resource!!!"));
+                .orElseThrow(() -> new UnAuthorizedOperationExcpetion("Unauthorized Access to Resource!!!"));
+
         Sort sort = sortDir.equalsIgnoreCase("asc") ? Sort.by(sortBy).ascending() : Sort.by(sortBy).descending();
         Pageable pageable = PageRequest.of(pageNumber, pageSize, sort);
-        Page<Comment> page_comment = commentRepository.findByUser(user, pageable);
-        List<CommentDTO> all_comments = page_comment.getContent().stream().map(CommentDTO::new).toList();
-        return new CommentPageResponse(all_comments,
-                page_comment.getNumber(),
-                page_comment.getSize(),
-                page_comment.getTotalElements(),
-                page_comment.getTotalPages(), page_comment.isLast());
+        Page<Comment> pageComment = commentRepository.findByUser(user, pageable);
 
+        List<CommentDTO> allComments = pageComment.getContent().stream().map(CommentDTO::new).toList();
+        return new CommentPageResponse(allComments,
+                pageComment.getNumber(),
+                pageComment.getSize(),
+                pageComment.getTotalElements(),
+                pageComment.getTotalPages(),
+                pageComment.isLast());
     }
 }
